# Speed-boost for close!(dh::MixedDofHandler)