Edges out a couple of anti-patterns in `close!(dh::MixedDofHandler)`. 

Example benchmark:
```
function construct_mixed_dof_handler(grid)
    mixed_dh = MixedDofHandler(grid)
    push!(mixed_dh, :f1, 1, Lagrange{2,RefCube,1}())
    push!(mixed_dh, :f2, 2, Lagrange{2,RefCube,1}())
    push!(mixed_dh, :f3, 1, Lagrange{2,RefCube,2}())
    close!(mixed_dh);
end
```
```
small_grid = generate_grid(Quadrilateral, (10, 10))
@btime construct_mixed_dof_handler($small_grid); 

# master: 196.791 μs (4493 allocations: 324.43 KiB)
# this PR: 52.959 μs (153 allocations: 186.04 KiB)
```
or for a bigger grid:
```
big_grid = generate_grid(Quadrilateral, (1000, 1000))
@btime construct_mixed_dof_handler($big_grid); 

# master: 3.821 s (70005918 allocations: 2.69 GiB)
# this PR: 1.229 s (454 allocations: 1.05 GiB)
```


_Sidenote_: Unfortunately the MixedDofHandler is still twice as slow as the DofHandler even with this PR, for comparison:
```
function construct_dof_handler(grid)
    dh = DofHandler(grid)
    push!(dh, :f1, 1, Lagrange{2,RefCube,1}())
    push!(dh, :f2, 2, Lagrange{2,RefCube,1}())
    push!(dh, :f3, 1, Lagrange{2,RefCube,2}())
    close!(dh)
end
@btime construct_dof_handler($small_grid) # 28.209 μs (140 allocations: 119.84 KiB)
@btime construct_dof_handler($big_grid) # 751.856 ms (341 allocations: 630.22 MiB)
```